add delegate
