+++ BruteForce/Sources/Scenes/Model/BruteForce.swift
@@ -1,26 +1,66 @@
 import UIKit
 
-func indexOf(character: Character, _ array: [String]) -> Int {
-    return array.firstIndex(of: String(character))!
-}
+//MARK: - Protocol
 
-func characterAt(index: Int, _ array: [String]) -> Character {
-    return index < array.count ? Character(array[index]) : Character("")
-}
-
-func generateBruteForce(_ string: String, fromArray array: [String]) -> String {
-    var str: String = string
+protocol UpdateInterface {
     
-    if str.count <= 0 {
-        str.append(characterAt(index: 0, array))
-    } else {
-        str.replace(at: str.count - 1,
-                    with: characterAt(index: (indexOf(character: str.last!, array) + 1) % array.count, array))
+    func updateInterface()
+}
+
+//MARK: - Class BruteForce
+
+class BruteForce {
+    
+    var delegate: UpdateInterface?
+    
+    var password = ""
+    
+    func generatePassword() -> String {
+        let characters = String().printable.map { String($0)}
         
-        if indexOf(character: str.last!, array) == 0 {
-            str = String(generateBruteForce(String(str.dropLast()), fromArray: array)) + String(str.last!)
+        for _ in 0..<3 {
+            password += characters.randomElement() ?? ""
         }
+        return password
     }
     
-    return str
+    func bruteForce(passwordToUnlock: String) {
+        
+        let ALLOWED_CHARACTERS: [String] = String().printable.map { String($0) }
+        
+        var password: String = ""
+        
+        while password != passwordToUnlock {
+            password = generateBruteForce(password, fromArray: ALLOWED_CHARACTERS)
+        }
+        
+        delegate?.updateInterface()
+        
+        print(password)
+    }
+    
+    func indexOf(character: Character, _ array: [String]) -> Int {
+        return array.firstIndex(of: String(character))!
+    }
+    
+    func characterAt(index: Int, _ array: [String]) -> Character {
+        return index < array.count ? Character(array[index]) : Character("")
+    }
+    
+    func generateBruteForce(_ string: String, fromArray array: [String]) -> String {
+        var str: String = string
+        
+        if str.count <= 0 {
+            str.append(characterAt(index: 0, array))
+        } else {
+            str.replace(at: str.count - 1,
+                        with: characterAt(index: (indexOf(character: str.last!, array) + 1) % array.count, array))
+            
+            if indexOf(character: str.last!, array) == 0 {
+                str = String(generateBruteForce(String(str.dropLast()), fromArray: array)) + String(str.last!)
+            }
+        }
+        
+        return str
+    }
 }
